pyproject.toml
[tool.poetry]
name = "farmers-market-backend"
version = "0.1.0"
description = ""
authors = ["Alexander Lubovsky <alexander.lubovsky@intexsoft.by>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.13.1"
Django = "^4.0.6"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
